ingestion_demo/bronze_ingestion/requirements.txt
pandas==2.2.2
psycopg2-binary==2.9.9
boto3==1.34.149
awswrangler==3.9.0